Fixing LDAPConnector timeout bug.
The LDAPConnector was set as static member with a timeout of 5000 seconds. Thus,
the connection would be closed after that time-period, and the service would need
to be restarted to allow for users to log-in. This commit removes using the
connection as a static member, and just instantiating a new connection on every
login with a try with resource.