Improve performance of caching layer.
The overall goal of this is to reduce time inside locks.
The caching layer should follow the pattern of optimistically
looking up the first layer and only locking the first layer
if it comes back null.  Then, change the second layer to
rely on Java's ConcurrentHashMap for handling concurrency there.

Bump version to 1.9.7.